Completed additing additional attributes to the responsetime report

+++ transfer-objects/src/main/java/org/sola/services/boundary/transferobjects/casemanagement/ResponseViewTO.java
@@ -44,7 +44,7 @@ public class ResponseViewTO extends AbstractBasicIdTO {
     
     private Integer totalTime;
     
-    private Integer averageTime;
+    private double averageTime;
     
     private Integer frequentDay;
     
@@ -100,11 +100,11 @@ public class ResponseViewTO extends AbstractBasicIdTO {
         this.stdDeviation = stdDeviation;
     }
 
-    public Integer getAverageTime() {
+    public double getAverageTime() {
         return averageTime;
     }
 
-    public void setAverageTime(Integer averageTime) {
+    public void setAverageTime(double averageTime) {
         this.averageTime = averageTime;
     }
